% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cv.R
\name{cv}
\alias{cv}
\title{Coefficient of variation}
\usage{
cv(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical indicating whether NA values should be stripped
before the computation proceeds. Default to \code{TRUE}.}
}
\value{
The coefficient of variation of \code{x}.
}
\description{
Coefficient of variation
}
\examples{
cv(1:10)
}
\author{
Antoine Bichat
}
