% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_hclust.R, R/aggregate_list.R,
%   R/mfa.R, R/rownormalize_tss.R, R/select_background.R, R/select_cv.R,
%   R/select_kruskal.R, R/select_wilcoxon.R, R/taxonomy.R
\name{required_pkgs.step_aggregate_hclust}
\alias{required_pkgs.step_aggregate_hclust}
\alias{required_pkgs.step_aggregate_list}
\alias{required_pkgs.step_mfa}
\alias{required_pkgs.step_rownormalize_tss}
\alias{required_pkgs.step_select_background}
\alias{required_pkgs.step_select_cv}
\alias{required_pkgs.step_select_kruskal}
\alias{required_pkgs.step_select_wilcoxon}
\alias{required_pkgs.step_taxonomy}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
\method{required_pkgs}{step_aggregate_hclust}(x, ...)

\method{required_pkgs}{step_aggregate_list}(x, ...)

\method{required_pkgs}{step_mfa}(x, ...)

\method{required_pkgs}{step_rownormalize_tss}(x, ...)

\method{required_pkgs}{step_select_background}(x, ...)

\method{required_pkgs}{step_select_cv}(x, ...)

\method{required_pkgs}{step_select_kruskal}(x, ...)

\method{required_pkgs}{step_select_wilcoxon}(x, ...)

\method{required_pkgs}{step_taxonomy}(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector
}
\description{
Recipe-adjacent packages always list themselves as a required package so that
the steps can function properly within parallel processing schemes.
}
\keyword{internal}
