% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimo-package.R
\docType{package}
\name{scimo-package}
\alias{scimo}
\alias{scimo-package}
\title{scimo: Extra Recipes Steps for Dealing with Omics Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Omics data (e.g. transcriptomics, proteomics, metagenomics...) offer a detailed and multi-dimensional perspective on the molecular components and interactions within complex biological (eco)systems. Analyzing these data requires adapted procedures, which are implemented as steps according to the 'recipes' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abichat/scimo}
  \item \url{https://abichat.github.io/scimo}
  \item Report bugs at \url{https://github.com/abichat/scimo/issues}
}

}
\author{
\strong{Maintainer}: Antoine BICHAT \email{antoine.bichat@proton.me} (\href{https://orcid.org/0000-0001-6599-7081}{ORCID})

Other contributors:
\itemize{
  \item Julie AUBERT \email{julie.aubert@inrae.fr} (\href{https://orcid.org/0000-0001-5203-5748}{ORCID}) [contributor]
}

}
\keyword{internal}
