% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{step_mfa}
\alias{step_mfa}
\alias{tidy.step_mfa}
\title{Dimension reduction step using multiple factor analysis}
\usage{
step_mfa(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  list_groups = NULL,
  type_var = rep("c", length(list_groups)),
  num_comp = 5,
  res = NULL,
  prefix = "mfa_",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("mfa")
)

\method{tidy}{step_mfa}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[=selections]{selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? By default, the new columns created by this step from
the original variables will be used as \code{predictors} in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{list_groups}{Named list specifying the groups of variables.}

\item{type_var}{A vector indicating the type of each group. \code{"c"} and \code{"s"}
are for quantitative variables (the difference is that for \code{"s"} variables
are scaled to unit variance), \code{"n"} for categorical variables \code{"m"} for
group of mixed variables and \code{"f"}for frequencies (from a contingency
table). By default, all variable are assumed \code{"c"}.}

\item{num_comp}{The number of components to retain as new predictors.}

\item{res}{This parameter is only produced after the recipe has been trained.}

\item{prefix}{A character string for the prefix of the resulting
new variables.}

\item{keep_original_cols}{A logical to keep the original variables in
the output. Defaults to \code{FALSE}.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}

\item{x}{A \code{step_mfa} object.}
}
\value{
An updated version of recipe with the new step added to the
sequence of any existing operations.
}
\description{
Project variables into a smaller space.
}
\examples{
\dontshow{if (rlang::is_installed("FactoMineR")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rec <-
  mtcars \%>\%
  transform(cyl = as.factor(cyl), vs = as.factor(vs),
            am = as.factor(am), gear = as.factor(gear),
            carb = as.factor(carb)) \%>\%
  tibble::as_tibble(rownames = "car") \%>\%
  recipe(car ~ ., data = .) \%>\%
  step_mfa(all_predictors(),
           list_groups = list(g1 = colnames(mtcars)[c(1, 3:7)],
                              g2 = colnames(mtcars)[c(2, 8:11)]),
           type_var = c("c", "n")) \%>\%
  prep()
rec
tidy(rec, 1)
bake(rec, new_data = NULL)
\dontshow{\}) # examplesIf}
}
\author{
Antoine Bichat
}
