% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{required_pkgs.step_taxonomy}
\alias{required_pkgs.step_taxonomy}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
\method{required_pkgs}{step_taxonomy}(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector
}
\description{
Recipe-adjacent packages always list themselves as a required package so that
the steps can function properly within parallel processing schemes.
}
\keyword{internal}
